Refactor duplicate pending voice image logic into a shared factory.
Moved `createPendingVoiceImage` and its helper functions out of
`CreateVoiceActivity.kt` and `DashboardPostDetailActivity.kt` into a
new singleton `VoiceImageFactory.kt`.

Fixes code health issue of duplicated creation and compression routines
for pending images.
