+++ app/src/main/java/org/ole/planet/myplanet/lite/dashboard/CreateVoiceActivity.kt
@@ -999,7 +999,7 @@ class CreateVoiceActivity : BaseActivity() {
 
     private suspend fun handleImageSelection(uri: Uri) {
         val pendingResult = withContext(Dispatchers.IO) {
-            runCatching { createPendingVoiceImage(uri) }
+            runCatching { VoiceImageFactory.createPendingVoiceImage(uri, contentResolver, cacheDir, ::generatePendingImageId) }
         }
         pendingResult.onSuccess { pending ->
             pendingImages[pending.id] = pending
@@ -1023,7 +1023,7 @@ class CreateVoiceActivity : BaseActivity() {
             editInitialImagePaths.map { path ->
                 async(Dispatchers.IO) {
                     semaphore.withPermit {
-                        runCatching { VoiceImageFetcher.fetchExistingImage(httpClient, cacheDir, sessionCookie, base, path, { generateImageFileName() }, { generatePendingImageId(it) }) }.getOrNull()
+                        runCatching { VoiceImageFetcher.fetchExistingImage(httpClient, cacheDir, sessionCookie, base, path, { VoiceImageFactory.generateImageFileName() }, { generatePendingImageId(it) }) }.getOrNull()
                     }
                 }
             }.awaitAll().filterNotNull()
@@ -1107,27 +1107,6 @@ class CreateVoiceActivity : BaseActivity() {
         return builder.toString() to seen
     }
 
-    private fun createPendingVoiceImage(uri: Uri): PendingVoiceImage {
-        val original = contentResolver.openInputStream(uri)?.use { input ->
-            BitmapFactory.decodeStream(input)
-        } ?: throw IllegalArgumentException("Unable to decode image stream")
-        val processed = prepareBitmapForWeb(original)
-        if (processed !== original) {
-            original.recycle()
-        }
-        val jpegBytes = compressBitmapToJpeg(processed)
-        if (!processed.isRecycled) {
-            processed.recycle()
-        }
-        val fileName = generateImageFileName()
-        val tempFile = File(cacheDir, fileName)
-        FileOutputStream(tempFile).use { output ->
-            output.write(jpegBytes)
-        }
-        val id = generatePendingImageId(fileName)
-        return PendingVoiceImage(id, fileName, tempFile, jpegBytes)
-    }
-
     private fun insertTemporaryImagePlaceholder(fileName: String) {
         val editable = createVoiceInput.text ?: return
         val placeholder = "![]($fileName)"
@@ -1143,28 +1122,6 @@ class CreateVoiceActivity : BaseActivity() {
         val updated = builder.toString()
         editable.replace(0, editable.length, updated)
         createVoiceInput.setSelection(updated.length)
-    }
-
-    private fun prepareBitmapForWeb(source: Bitmap): Bitmap {
-        val maxSide = max(source.width, source.height)
-        if (maxSide <= MAX_IMAGE_DIMENSION) {
-            return source
-        }
-        val scale = MAX_IMAGE_DIMENSION.toFloat() / maxSide.toFloat()
-        val width = (source.width * scale).roundToInt().coerceAtLeast(1)
-        val height = (source.height * scale).roundToInt().coerceAtLeast(1)
-        return Bitmap.createScaledBitmap(source, width, height, true)
-    }
-
-    private fun compressBitmapToJpeg(bitmap: Bitmap): ByteArray {
-        val output = ByteArrayOutputStream()
-        bitmap.compress(Bitmap.CompressFormat.JPEG, JPEG_QUALITY, output)
-        return output.toByteArray()
-    }
-
-    private fun generateImageFileName(): String {
-        val formatter = SimpleDateFormat("'post'yyyyMMddHHmmssSSS", Locale.US)
-        return formatter.format(Date()) + ".jpg"
     }
 
     private fun generatePendingImageId(baseName: String): String {
@@ -1594,8 +1551,6 @@ class CreateVoiceActivity : BaseActivity() {
     companion object {
         private const val PREVIEW_DEBOUNCE_MS = 150L
         private const val MAX_HEADING_LEVEL = 6
-        private const val MAX_IMAGE_DIMENSION = 1280
-        private const val JPEG_QUALITY = 85
         private val NUMBERED_LIST_REGEX = Regex("^(\\d+)\\.\\s*(.*)$")
         private const val KEY_DEVICE_ANDROID_ID = "device_android_id"
         private const val KEY_DEVICE_CUSTOM_DEVICE_NAME = "device_custom_device_name"
