+++ app/src/main/java/org/ole/planet/myplanet/lite/dashboard/DashboardPostDetailActivity.kt
@@ -799,7 +799,7 @@ class DashboardPostDetailActivity : AppCompatActivity() {
 
     private suspend fun handleReplyImageSelection(uri: Uri) {
         val pendingResult = withContext(Dispatchers.IO) {
-            runCatching { createPendingVoiceImage(uri) }
+            runCatching { VoiceImageFactory.createPendingVoiceImage(uri, contentResolver, cacheDir, ::generatePendingImageId) }
         }
         pendingResult.onSuccess { pending ->
             pendingReplyImages[pending.id] = pending
@@ -1170,7 +1170,7 @@ class DashboardPostDetailActivity : AppCompatActivity() {
         }
         val loaded = coroutineScope {
             comment.imagePaths.map { path ->
-                async(Dispatchers.IO) { VoiceImageFetcher.fetchExistingImage(httpClient, cacheDir, sessionCookie, base, path, { generateImageFileName() }, { generatePendingImageId(it) }) }
+                async(Dispatchers.IO) { VoiceImageFetcher.fetchExistingImage(httpClient, cacheDir, sessionCookie, base, path, { VoiceImageFactory.generateImageFileName() }, { generatePendingImageId(it) }) }
             }.awaitAll().filterNotNull().toMutableList()
         }
         if (loaded.isEmpty()) {
@@ -1329,49 +1329,6 @@ class DashboardPostDetailActivity : AppCompatActivity() {
             else -> null
         }
         return base
-    }
-
-    private fun createPendingVoiceImage(uri: Uri): PendingVoiceImage {
-        val original = contentResolver.openInputStream(uri)?.use { input ->
-            BitmapFactory.decodeStream(input)
-        } ?: throw IllegalArgumentException("Unable to decode image stream")
-        val processed = prepareBitmapForWeb(original)
-        if (processed !== original) {
-            original.recycle()
-        }
-        val jpegBytes = compressBitmapToJpeg(processed)
-        if (!processed.isRecycled) {
-            processed.recycle()
-        }
-        val fileName = generateImageFileName()
-        val tempFile = File(cacheDir, fileName)
-        FileOutputStream(tempFile).use { output ->
-            output.write(jpegBytes)
-        }
-        val id = generatePendingImageId(fileName)
-        return PendingVoiceImage(id, fileName, tempFile, jpegBytes)
-    }
-
-    private fun prepareBitmapForWeb(source: Bitmap): Bitmap {
-        val maxSide = max(source.width, source.height)
-        if (maxSide <= MAX_IMAGE_DIMENSION) {
-            return source
-        }
-        val scale = MAX_IMAGE_DIMENSION.toFloat() / maxSide.toFloat()
-        val width = (source.width * scale).toInt().coerceAtLeast(1)
-        val height = (source.height * scale).toInt().coerceAtLeast(1)
-        return Bitmap.createScaledBitmap(source, width, height, true)
-    }
-
-    private fun compressBitmapToJpeg(bitmap: Bitmap): ByteArray {
-        val output = ByteArrayOutputStream()
-        bitmap.compress(Bitmap.CompressFormat.JPEG, JPEG_QUALITY, output)
-        return output.toByteArray()
-    }
-
-    private fun generateImageFileName(): String {
-        val formatter = SimpleDateFormat("'post'yyyyMMddHHmmssSSS", Locale.US)
-        return formatter.format(Date()) + ".jpg"
     }
 
     private fun generatePendingImageId(baseName: String): String {
@@ -1987,8 +1944,6 @@ class DashboardPostDetailActivity : AppCompatActivity() {
         private const val DAY_MILLIS = 24 * HOUR_MILLIS
         private const val MONTH_MILLIS = 30 * DAY_MILLIS
         private const val YEAR_MILLIS = 12 * MONTH_MILLIS
-        private const val JPEG_QUALITY = 85
-        private const val MAX_IMAGE_DIMENSION = 1280
         private const val PRIVATE_FOR_COMMUNITY = "community"
         private const val PREFS_NAME = "PREFS"
         private const val KEY_DEVICE_ANDROID_ID = "device_android_id"
